move the dump object outside of the recovering
