+++ wallet/src/lib.rs
@@ -2,12 +2,12 @@ mod blockchain;
 mod keygen;
 mod password;
 mod recovering;
-mod transaction;
+pub mod transaction;
 
 pub use self::{
     blockchain::Settings,
     password::{Password, ScrubbedBytes},
-    recovering::{Dump, RecoveringDaedalus, RecoveringIcarus, RecoveryBuilder, RecoveryError},
+    recovering::{RecoveringDaedalus, RecoveringIcarus, RecoveryBuilder, RecoveryError},
 };
 use hdkeygen::account::{Account, AccountId};
